feat: add api restcall that is commented out until cors error fixed on backend. refactor: move socket routing to utils folder.
