+++ utils/apiCalls.jsx
@@ -1,0 +1,4 @@
+export async function getMessage(){
+    const data = await fetch('http://localhost:5000/all_users')
+    console.log(data)
+}
